fixed so users can't see the 'delete this post' message on profiles not their own. also getting a weird bug where redirects aren't working locally, deploying to see if live site has the same problem

+++ views/v_users_profile.php
@@ -20,20 +20,21 @@
             <h3>my posts</h3>
 
             <?php foreach($user_posts as $user_post): ?>
-	        <p>
+                <p>
+    		        <span class='timestamp'>
+	                   <?=date('m/d/y | g:ia', $user_post['created'])?>:<br>
+		            </span>
 
-		<span class='timestamp'>
-	            <?=date('m/d/y | g:ia', $user_post['created'])?>:<br>
-		</span>
+	                <?=$user_post['content']?><br>
 
-	        <?=$user_post['content']?><br>
-
-		<a href='/posts/delete/<?=$user_post['post_id']?>'>delete this post</a>
-	        </p>
-	    <?php endforeach; ?>
+                    <?php if($logged_in_user_id == $profile['user_id']): ?>
+    		            <a href='/posts/delete/<?=$user_post['post_id']?>'>delete this post</a>
+                    <?php endif; ?>
+	            </p>
+	        <?php endforeach; ?>
 
         </div>
 
     <?php endif; ?>
 
-<? endif; ?>
+<?php endif; ?>
